Platform change + updates
Deleted Vaadin prohect due to no support for google maps. It would have been too long and too difficult to implement google maps.
So I just rewrote the math for calculating sunset sunrise in Python and implemented PHP to run that code in a shell.
The map is displayed mostly with JS.
Many things missing but they are coming soon.